Move OCP MachineSetDeployer to it's own package
The code has been spread across different places, but should really be
in a separate package as it's not coupled with AWS and in fact is
agnostic to the cloud.
